requirements.txt
fastapi==0.115.14
uvicorn==0.34.3
psycopg2-binary==2.9.9
pydantic==2.11.7
sqlalchemy==2.0.41
alembic==1.13.1
openai==1.54.0
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
